+++ src/main/kotlin/no/nav/familie/ef/sak/selvstendig/NæringsinntektKontrollScheduler.kt
@@ -5,6 +5,7 @@ import no.nav.familie.ef.sak.infrastruktur.featuretoggle.Toggle
 import org.springframework.context.annotation.Profile
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
+import java.time.YearMonth
 
 @Profile("!integrasjonstest")
 @Service
@@ -15,7 +16,7 @@ class NæringsinntektKontrollScheduler(
     @Scheduled(initialDelay = 60 * 1000L, fixedDelay = 365 * 24 * 60 * 60 * 1000L) // Kjører ved oppstart av app
     fun sjekkNæringsinntekt() {
         if (featureToggleService.isEnabled(Toggle.KONTROLLER_NÆRINGSINNTEKT)) {
-            næringsinntektKontrollService.kontrollerInntektForSelvstendigNæringsdrivende()
+            næringsinntektKontrollService.kontrollerInntektForSelvstendigNæringsdrivende(YearMonth.now().year - 1)
         }
     }
 }
